PROCJAM UPDATE: Traps, Making Camp!
This update to my wonderful robochild adds two main new threads, to the Ordeal, Allies, and Tests stages of the Hero's Journey.

The first is camping: The party will now make camp with a descriptor, and perform a generated task such as cooking or meditating or singing songs. It's a wholesome touch!

Thes second is dodging traps: This will start a dungeon-delving thread that will expand the Tests, Enemies, and Ordeal stages, and comes with a new little name generator for the McGuffin.

I'll write more of the dungeon scenes over the next month, and then include the new arrays within the second-person tweets! I love them and they get the most attention, so they should be expanded with the rest of the bot.