Update: A Play Within A Play
Hot on the heels of the **incredible** Station Eleven, which features a story within a story, scenes from the world-famous Hamlet! In the final episode, Shakespeare's words become a conduit for emotions that some of the main characters have, but can't fully process. It was one of at least 5 scenes that brought me to tears!!!
So I created a small play name generator under #play2# (which I messily nested into #play# with '' around it to make it easier to read, I know, it was an afterthought). Some of these are based on famous works and the generator will, on occasion, reconstitute a famous name (like 'The Normal Heart'). More vocabulary is added with #subs# for substances and punny playwright names under #playw#, so plenty to discover!
Enjoy!
- Ewan